area admin - não permitir apagar anos letivo com escolas associadas

+++ admin/includes/ano_letivo.php
@@ -1,11 +1,19 @@
 <?php
 
 	if($_GET['delete']==1){
-		$res = mysql_query("DELETE FROM `dbo_tab_ano_letivo` WHERE ID_ANO_LETIVO = $_GET[id]");
+		$search = mysql_query("SELECT * from dbo_tab_escolas where ID_ANO_LETIVO = $_GET[id] limit 1");
+		$num_rows = mysql_num_rows($search);
+
+		if ($num_rows == 0) {
+			$res = mysql_query("DELETE FROM `dbo_tab_ano_letivo` WHERE ID_ANO_LETIVO = $_GET[id]");
+		}else if ($num_rows == 1 && !empty($_GET['id'])){
+			echo '<div class="alert alert-warning">
+				<a href="#" class="close" data-dismiss="alert" aria-label="close">&times;</a>
+			  <a data-dismiss="alert"><strong>Atenção!</strong> Não é possível apagar um ano letivo que já tenha escolas registadas!</a>
+			</div>';
+		}
 	}
-
 ?>
-
 <div class="row">
 	<div class="col-xs-12">
 	  <div class="box">
@@ -77,7 +85,7 @@
 			  "targets": 'no-sort',
 			  "orderable": false,
 			}],
-			"order": [[ 0, "asc" ]]
+			"order": [[ 0, "desc" ]]
 		});
 	});
 </script>
